Starting to split the character class into an .h .cpp file.

+++ Driver.cpp
@@ -133,8 +133,8 @@ class Character
 	private: 
                 
         string name;
-        double healthPoints;
-        double attackStat;
+        int healthPoints;
+        int attackStat;
         int speed;
         string* arrayOfitems;
         int numItems;
